refactor: server._publicFiles
